% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{add_prefix_column ...}
\alias{add_prefix_column ...}
\alias{add_prefix_column}
\title{...}
\usage{
add_prefix_column(time_series)
}
\arguments{
\item{time_series}{...}
}
\value{
....
}
\description{
...
}
