% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_metrics.R
\name{area_q2.matrix}
\alias{area_q2.matrix}
\title{...}
\usage{
\method{area_q2}{matrix}(timeseries)
}
\arguments{
\item{timeseries}{...}
}
\value{
...
}
\description{
Area of the closed shape over the second quadrant
}
