% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_metrics.R
\name{area_q3.list}
\alias{area_q3.list}
\title{...}
\usage{
\method{area_q3}{list}(list_of_polygons)
}
\arguments{
\item{list_of_polygons}{...}
}
\value{
...
}
\description{
Area of the closed shape over the third quadrant.
}
