% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_metrics.R
\name{area_q3.numeric}
\alias{area_q3.numeric}
\title{...}
\usage{
\method{area_q3}{numeric}(timeseries)
}
\arguments{
\item{timeseries}{...}
}
\value{
...
}
\description{
Area of the closed shape over the third quadrant.
}
