% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_metrics.R
\name{area_ts.sfc_POLYGON}
\alias{area_ts.sfc_POLYGON}
\alias{area_ts.matrix}
\title{...}
\usage{
\method{area_ts}{matrix}(polygon)
}
\arguments{
\item{polygon}{...}
}
\value{
...
}
\description{
Area of the closed shape
}
