% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{basics_metrics ...}
\alias{basics_metrics ...}
\alias{basics_metrics}
\title{...}
\usage{
basics_metrics(sits_timeseries)
}
\arguments{
\item{sits_timeseries}{...}
}
\value{
....
}
\description{
...
}
