% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch22_metrics.R
\name{catch22_metrics}
\alias{catch22_metrics}
\alias{dn_histmode_5}
\title{Catch22 metrics}
\usage{
dn_histmode_5(timeseries)
}
\arguments{
\item{timeseries}{a \code{numeric} or \code{matrix} object where the
columns is the point in time.}
}
\value{
a \code{numeric} vector for each metric in each time series
}
\description{
The package sitsfeats uses Rcatch22 package to provide catch22 metrics:
\itemize{
  \item ...
  \item ...
 }
}
\examples{
data("timeseries")
mean_values <- sitsfeats::dn_histmode_5(timeseries)

}
