% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_metrics.R
\name{create_polygon_v2}
\alias{create_polygon_v2}
\title{...}
\usage{
create_polygon_v2(timeseries)
}
\arguments{
\item{timeseries}{...}
}
\value{
a polygon object from \code{sf} class
}
\description{
...
}
