% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_metrics.R
\name{csi.matrix}
\alias{csi.matrix}
\title{...}
\usage{
\method{csi}{matrix}(timeseries)
}
\arguments{
\item{timeseries}{...}
}
\value{
...
}
\description{
This is a dimensionless quantitative measure of morphology
}
