% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_metrics.R
\name{csi.POLYGON}
\alias{csi.POLYGON}
\alias{csi.XY}
\title{...}
\usage{
\method{csi}{XY}(polygon)
}
\arguments{
\item{polygon}{...}
}
\value{
...
}
\description{
This is a dimensionless quantitative measure of morphology
}
