% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.verify_timeseries}
\alias{.verify_timeseries}
\title{Verify time series}
\usage{
.verify_timeseries(timeseries)
}
\arguments{
\item{timeseries}{a \code{numeric} or \code{matrix} object where the
columns is the point in time.}
}
\value{
a \code{matrix} from time series.
}
\description{
Verify time series
}
