% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_metrics.R
\name{ecc_metric.numeric}
\alias{ecc_metric.numeric}
\title{...}
\usage{
\method{ecc_metric}{numeric}(timeseries)
}
\arguments{
\item{timeseries}{...}
}
\value{
...
}
\description{
Return values close to 0 if the shape is a circle and 1
}
