% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_metrics.R
\name{ecc_metric.POLYGON}
\alias{ecc_metric.POLYGON}
\alias{ecc_metric.XY}
\title{...}
\usage{
\method{ecc_metric}{XY}(polygon)
}
\arguments{
\item{polygon}{...}
}
\value{
...
}
\description{
Return values close to 0 if the shape is a circle and 1
}
