% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{entropy_filter}
\alias{entropy_filter}
\title{Entropy filter uses information gain and other methods to determine
 more accurately feature selection.}
\usage{
entropy_filter(list_metrics, y, type = "infogain")
}
\arguments{
\item{y}{label samples}

\item{type}{a charactecter corresponding the method to select features}

\item{x}{time series matrix}

\item{ntree}{number of tree}
}
\value{
a time series with subset selected
}
\description{
Entropy filter uses information gain and other methods to determine
 more accurately feature selection.
}
