% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{extracting_metrics_basics ...}
\alias{extracting_metrics_basics ...}
\alias{extracting_metrics_basics}
\title{...}
\usage{
extracting_metrics_basics(sits_timeseries)
}
\arguments{
\item{sits_timeseries}{...}
}
\value{
....
}
\description{
...
}
