% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_all_areas}
\alias{get_all_areas}
\title{Get all areas from a closed polar polygon}
\usage{
get_all_areas(timeseries)
}
\arguments{
\item{timeseries}{a \code{numeric} or \code{matrix} object where the
columns is the point in time.}
}
\value{
a \code{matrix} with each area from each quadrant.
}
\description{
The standard deviation of the areas per season
}
