% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_metrics.R
\name{gyration_radius.numeric}
\alias{gyration_radius.numeric}
\title{...}
\usage{
\method{gyration_radius}{numeric}(timeseries)
}
\arguments{
\item{timeseries}{...}
}
\value{
...
}
\description{
Equals the average distance between each point inside
}
