% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_metrics.R
\name{gyration_radius}
\alias{gyration_radius}
\title{...}
\usage{
gyration_radius(type)
}
\arguments{
\item{type}{...}
}
\value{
...
}
\description{
Equals the average distance between each point inside
}
