% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_metrics.R
\name{gyration_radius.sfc_POLYGON}
\alias{gyration_radius.sfc_POLYGON}
\title{...}
\usage{
\method{gyration_radius}{sfc_POLYGON}(polygon)
}
\arguments{
\item{polygon}{...}
}
\value{
...
}
\description{
Equals the average distance between each point inside
}
