% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_metrics.R
\name{polar_balance.numeric}
\alias{polar_balance.numeric}
\title{...}
\usage{
\method{polar_balance}{numeric}(timeseries)
}
\arguments{
\item{timeseries}{...}
}
\value{
...
}
\description{
The standard deviation of the areas per season
}
