% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R, R/polar_metrics.R
\name{polar_plot}
\alias{polar_plot}
\title{...}
\usage{
polar_plot(timeseries)

polar_plot(timeseries)
}
\arguments{
\item{timeseries}{...}
}
\value{
a polygon object from \code{sf} class

a polygon object from \code{sf} class
}
\description{
...

...
}
