\name{sitsfeats-package}
\alias{sitsfeats-package}
\alias{sitsfeats}
\docType{package}
\title{
\packageTitle{sitsfeats}
}
\description{
\packageDescription{sitsfeats}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sitsfeats}
\packageIndices{sitsfeats}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{sitsfeats}

Maintainer: \packageMaintainer{sitsfeats}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
