% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{wrapper_filter}
\alias{wrapper_filter}
\title{wrapper filter uses a regularized random forest}
\usage{
wrapper_filter(list_metrics, y, flagReg = 1, ntree = 500)
}
\arguments{
\item{y}{label samples}

\item{flagReg}{if 1 is generated a RRF tree, otherwise a normal RF}

\item{ntree}{number of tree}

\item{x}{time series matrix}
}
\value{
a time series with subset selected
}
\description{
wrapper filter uses a regularized random forest
}
