% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_polygon_geos}
\alias{create_polygon_geos}
\title{...}
\usage{
create_polygon_geos(timeseries)
}
\arguments{
\item{timeseries}{...}
}
\value{
a polygon object from \code{sf} class
}
\description{
...
}
