% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_metrics.R
\name{polar_metrics}
\alias{polar_metrics}
\title{...}
\usage{
polar_metrics(timeseries, metrics = c())
}
\arguments{
\item{timeseries}{...}

\item{metrics}{...}
}
\value{
...
}
\description{
...
}
\note{
é necessária a instação do pacote grDevices para o método chull
}
