% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{polar_plot}
\alias{polar_plot}
\title{Creates a polar plot based on areas ...}
\usage{
polar_plot(timeseries)
}
\arguments{
\item{timeseries}{a \code{numeric} or \code{matrix} object where the
columns is the point in time.}
}
\value{
a polygon object from \code{sf} class
}
\description{
Creates a polar plot based on areas ...
}
